pyproject.toml
[project]
name = "api-crm"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "guilherme", email = "vinespedro55@gmail.com"},
]
dependencies = [
    "pylint>=3.0.3",
    "flake8>=7.0.0",
    "pre-commit>=3.7.1",
    "ruff>=0.6.5",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "email-validator>=2.2.0",
    "bcrypt>=4.2.0",
    "pytest-asyncio>=0.24.0",
    "pytest-django>=4.9.0",
    "pytest-env>=1.1.5",
    "psycopg2-binary>=2.9.9",
    "django>=5.1.1",
    "django-ninja>=1.3.0",
    "django-cors-headers>=4.4.0",
    "pyjwt>=2.9.0",
    "django-cockroachdb>=5.1",
    "django-storages>=1.14.4",
    "boto3>=1.36.20",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"


[tool.pytest.ini_options]
testpaths = ["tests", "src"]
python_files = ["test_*.py", "*_test.py"]
asyncio_default_fixture_loop_scope = "function"
DJANGO_SETTINGS_MODULE = "src.framework.settings"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pylint>=3.0.3",
    "autopep8>=2.0.4",
    "pytest>=8.0.1",
]


[tool.ruff]
src = ["src"]
line-length = 88
exclude = ["*/migrations/*"]


[tool.ruff.lint]
ignore = ["N818", "EXE002"]
select = [
    "F",   # Pyflakes
    "E",   # pycodestyle - Errors
    "W",   # pycodestyle - Warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "YTT", # flake8-2020
    "COM", # flake8-commas
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "EM",  # flake8-errmsg
    "EXE", # flake8-executable
    "FA",  # flake8-future-annotations
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "G",   # flake8-logging-format
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "INT", # flake8-gettext
     "PL",  # Pylint
]
